fix: byte、short 运算后会变为 int 的问题

+++ src/main/java/org/jcnc/snow/vm/commands/arithmetic/byte8/BMulCommand.java
@@ -50,7 +50,7 @@ public class BMulCommand implements Command {
         byte a = (byte) operandStack.pop();
 
         // Perform the multiplication and push the result back onto the stack
-        operandStack.push(a * b);
+        operandStack.push((byte)(a * b));
 
         // Return the updated program counter (next instruction)
         return currentPC + 1;
